refactor(ui): update data storage
