+++ app/src/main/java/org/flyve/mdm/agent/ui/EnrollmentActivity.java
@@ -55,7 +55,7 @@ import android.widget.TextView;
 import org.flyve.mdm.agent.R;
 import org.flyve.mdm.agent.core.enrollment.Enrollment;
 import org.flyve.mdm.agent.core.enrollment.EnrollmentPresenter;
-import org.flyve.mdm.agent.core.user.UserData;
+import org.flyve.mdm.agent.data.UserData;
 import org.flyve.mdm.agent.utils.FlyveLog;
 import org.flyve.mdm.agent.utils.Helpers;
 import org.flyve.mdm.agent.utils.MultipleEditText;
