+++ app/src/main/java/org/flyve/mdm/agent/ui/StartEnrollmentActivity.java
@@ -41,8 +41,8 @@ import android.widget.TextView;
 
 import org.flyve.mdm.agent.R;
 import org.flyve.mdm.agent.core.enrollment.EnrollmentHelper;
-import org.flyve.mdm.agent.core.supervisor.SupervisorData;
-import org.flyve.mdm.agent.data.DataStorage;
+import org.flyve.mdm.agent.data.MqttData;
+import org.flyve.mdm.agent.data.SupervisorData;
 import org.flyve.mdm.agent.utils.FlyveLog;
 import org.flyve.mdm.agent.utils.Helpers;
 
@@ -67,7 +67,7 @@ public class StartEnrollmentActivity extends Activity {
 
         broadcastClose();
 
-        DataStorage cache = new DataStorage( StartEnrollmentActivity.this );
+        MqttData cache = new MqttData( StartEnrollmentActivity.this );
         // if broker is on cache open the main activity
         String broker = cache.getBroker();
         if(broker != null) {
